changed token fetch interval from 30 minutes to 120 minutes
Google Appengine Channel tokens are valid for 2 hours. Once a token is
created it, GAE immediately charges all CPU cycles for a 2hour
persistent connection. By having the token refresh on 30 minutes we are
actually wasting a lot of cpu cycles.

With this change tokens will only be refreshed after 2 hours when all
cycles that were billed at creation time are consumed.